+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Workout Tracker",
+  "name": "CS 411 Workout Tracker",
   "icons": [
     {
       "src": "favicon.ico",
